Rearange code
Split the code up into more logical related components.

Move josh-filter.rs to josh-cli crate
Those two make a lot of sense to be installed together.
Also we're considering to use the "josh-filter" crate name
to extract parts of josh-core.
