+++ Cargo.toml
@@ -5,7 +5,6 @@ members = [
     "hyper-reverse-proxy",
     "josh-core",
     "josh-cli",
-    "josh-filter",
     "josh-graphql",
     "josh-proxy",
     "josh-rpc",
