+++ josh-core/src/cache/mod.rs
@@ -1,0 +1,9 @@
+mod cache;
+pub mod notes;
+pub mod sled;
+pub mod stack;
+
+pub use cache::*;
+pub use notes::NotesCacheBackend;
+pub use sled::{SledCacheBackend, sled_load, sled_open_josh_trees, sled_print_stats};
+pub use stack::CacheStack;
