+++ josh-core/src/filter/mod.rs
@@ -1,13 +1,13 @@
 use super::*;
-use pest::Parser;
 use std::path::Path;
 use std::sync::LazyLock;
-mod op;
-mod opt;
-mod parse;
+pub(crate) mod op;
+pub(crate) mod opt;
 pub mod persist;
 pub mod text;
 pub mod tree;
+
+use crate::flang::parse;
 
 use op::{LazyRef, Op};
 
@@ -15,6 +15,7 @@ pub use persist::as_tree;
 pub use persist::from_tree;
 pub(crate) use persist::{to_filter, to_op};
 
+pub use crate::flang::{pretty, spec};
 pub use opt::invert;
 pub use parse::get_comments;
 pub use parse::parse;
@@ -26,7 +27,7 @@ static ANCESTORS: LazyLock<
 
 /// Match-all regex pattern used as the default for Op::Message when no regex is specified.
 /// The pattern `(?s)^.*$` matches any string (including newlines) from start to end.
-static MESSAGE_MATCH_ALL_REGEX: LazyLock<regex::Regex> =
+pub(crate) static MESSAGE_MATCH_ALL_REGEX: LazyLock<regex::Regex> =
     LazyLock::new(|| regex::Regex::new("(?s)^.*$").unwrap());
 
 /// Filters are represented as `git2::Oid`, however they are not ever stored
@@ -60,10 +61,12 @@ impl Filter {
     pub fn id(&self) -> git2::Oid {
         self.0
     }
-}
 
-pub fn sequence_number() -> Filter {
-    Filter(git2::Oid::zero())
+    /// Create a Filter from an Oid. This is primarily used for special filters
+    /// like sequence_number that don't correspond to a normal Op variant.
+    pub(crate) fn from_oid(oid: git2::Oid) -> Filter {
+        Filter(oid)
+    }
 }
 
 impl std::fmt::Debug for Filter {
@@ -201,128 +204,7 @@ impl<'a> Rewrite<'a> {
     }
 }
 
-pub fn nop() -> Filter {
-    to_filter(Op::Nop)
-}
-
-pub fn empty() -> Filter {
-    to_filter(Op::Empty)
-}
-
-pub fn message(m: &str) -> Filter {
-    to_filter(Op::Message(m.to_string(), MESSAGE_MATCH_ALL_REGEX.clone()))
-}
-
-pub fn file(path: impl Into<std::path::PathBuf>) -> Filter {
-    let p = path.into();
-    to_filter(Op::File(p.clone(), p))
-}
-
-pub fn hook(h: &str) -> Filter {
-    to_filter(Op::Hook(h.to_string()))
-}
-
-pub fn squash(ids: Option<&[(git2::Oid, Filter)]>) -> Filter {
-    if let Some(ids) = ids {
-        to_filter(Op::Squash(Some(
-            ids.iter()
-                .map(|(x, y)| (LazyRef::Resolved(*x), *y))
-                .collect(),
-        )))
-    } else {
-        to_filter(Op::Squash(None))
-    }
-}
-
-/// Pretty print the filter on multiple lines with initial indentation level.
-/// Nested filters will be indented with additional 4 spaces per nesting level.
-pub fn pretty(filter: Filter, indent: usize) -> String {
-    let filter = opt::simplify(filter);
-
-    if let Op::Compose(filters) = to_op(filter) {
-        if indent == 0 {
-            let i = format!("\n{}", " ".repeat(indent));
-            return filters
-                .iter()
-                .map(|x| pretty2(&to_op(*x), indent + 4, true))
-                .collect::<Vec<_>>()
-                .join(&i);
-        }
-    }
-    pretty2(&to_op(filter), indent, true)
-}
-
-fn pretty2(op: &Op, indent: usize, compose: bool) -> String {
-    let ff = |filters: &Vec<_>, n, ind| {
-        let ind2 = std::cmp::max(ind, 4);
-        let i = format!("\n{}", " ".repeat(ind2));
-        let joined = filters
-            .iter()
-            .map(|x| pretty2(&to_op(*x), ind + 4, true))
-            .collect::<Vec<_>>()
-            .join(&i);
-
-        format!(
-            ":{}[{}{}{}]",
-            n,
-            &i,
-            joined,
-            &format!("\n{}", " ".repeat(ind2 - 4))
-        )
-    };
-    match op {
-        Op::Compose(filters) => ff(filters, "", indent),
-        Op::Subtract(af, bf) => ff(&vec![*af, *bf], "subtract", indent + 4),
-        Op::Exclude(bf) => match to_op(*bf) {
-            Op::Compose(filters) => ff(&filters, "exclude", indent),
-            b => format!(":exclude[{}]", pretty2(&b, indent, false)),
-        },
-        Op::Pin(filter) => match to_op(*filter) {
-            Op::Compose(filters) => ff(&filters, "pin", indent),
-            b => format!(":pin[{}]", pretty2(&b, indent, false)),
-        },
-        Op::Chain(a, b) => match (to_op(*a), to_op(*b)) {
-            (Op::Subdir(p1), Op::Prefix(p2)) if p1 == p2 => {
-                format!("::{}/", parse::quote_if(&p1.to_string_lossy()))
-            }
-            (a, Op::Prefix(p)) if compose => {
-                format!(
-                    "{} = {}",
-                    parse::quote_if(&p.to_string_lossy()),
-                    pretty2(&a, indent, false)
-                )
-            }
-            (a, b) => format!(
-                "{}{}",
-                pretty2(&a, indent, false),
-                pretty2(&b, indent, false)
-            ),
-        },
-        Op::RegexReplace(replacements) => {
-            let v = replacements
-                .iter()
-                .map(|(regex, r)| {
-                    format!(
-                        "{}{}:{}",
-                        " ".repeat(indent),
-                        parse::quote(&regex.to_string()),
-                        parse::quote(r)
-                    )
-                })
-                .collect::<Vec<_>>();
-            format!(":replace(\n{}\n)", v.join("\n"))
-        }
-        Op::Squash(Some(ids)) => {
-            let mut v = ids
-                .iter()
-                .map(|(oid, f)| format!("{}{}{}", " ".repeat(indent), &oid.to_string(), spec(*f)))
-                .collect::<Vec<_>>();
-            v.sort();
-            format!(":squash(\n{}\n)", v.join("\n"))
-        }
-        _ => spec2(op),
-    }
-}
+pub use crate::build::{chain, compose, empty, file, hook, message, nop, sequence_number, squash};
 
 pub fn lazy_refs(filter: Filter) -> Vec<String> {
     lazy_refs2(&to_op(filter))
@@ -451,149 +333,6 @@ fn resolve_refs2(refs: &std::collections::HashMap<String, git2::Oid>, op: &Op) -
             Op::Squash(Some(lr))
         }
         _ => op.clone(),
-    }
-}
-
-/// Compact, single line string representation of a filter so that `parse(spec(F)) == F`
-/// Note that this is will not be the best human readable representation. For that see `pretty(...)`
-pub fn spec(filter: Filter) -> String {
-    if filter == sequence_number() {
-        return "sequence_number".to_string();
-    }
-    let filter = opt::simplify(filter);
-    spec2(&to_op(filter))
-}
-
-fn spec2(op: &Op) -> String {
-    match op {
-        Op::Compose(filters) => {
-            format!(
-                ":[{}]",
-                filters
-                    .iter()
-                    .map(|x| spec(*x))
-                    .collect::<Vec<_>>()
-                    .join(",")
-            )
-        }
-        Op::Subtract(a, b) => {
-            format!(":subtract[{},{}]", spec(*a), spec(*b))
-        }
-        Op::Exclude(b) => {
-            format!(":exclude[{}]", spec(*b))
-        }
-        Op::Pin(filter) => {
-            format!(":pin[{}]", spec(*filter))
-        }
-        Op::Rev(filters) => {
-            let mut v = filters
-                .iter()
-                .map(|(k, v)| format!("{}{}", k.to_string(), spec(*v)))
-                .collect::<Vec<_>>();
-            v.sort();
-            format!(":rev({})", v.join(","))
-        }
-        Op::Workspace(path) => {
-            format!(":workspace={}", parse::quote_if(&path.to_string_lossy()))
-        }
-        #[cfg(feature = "incubating")]
-        Op::Lookup(path) => {
-            format!(":lookup={}", parse::quote_if(&path.to_string_lossy()))
-        }
-        #[cfg(feature = "incubating")]
-        Op::Lookup2(oid) => {
-            format!(":lookup2={}", oid.to_string())
-        }
-        Op::Stored(path) => {
-            format!(":+{}", parse::quote_if(&path.to_string_lossy()))
-        }
-        Op::RegexReplace(replacements) => {
-            let v = replacements
-                .iter()
-                .map(|(regex, r)| {
-                    format!("{}:{}", parse::quote(&regex.to_string()), parse::quote(r))
-                })
-                .collect::<Vec<_>>();
-            format!(":replace({})", v.join(","))
-        }
-
-        Op::Chain(a, b) => match (to_op(*a), to_op(*b)) {
-            (Op::Subdir(p1), Op::Prefix(p2)) if p1 == p2 => {
-                format!("::{}/", parse::quote_if(&p1.to_string_lossy()))
-            }
-            (a, b) => format!("{}{}", spec2(&a), spec2(&b)),
-        },
-
-        Op::Nop => ":/".to_string(),
-        Op::Empty => ":empty".to_string(),
-        Op::Paths => ":PATHS".to_string(),
-        Op::Invert => ":INVERT".to_string(),
-        Op::Index => ":INDEX".to_string(),
-        Op::Fold => ":FOLD".to_string(),
-        Op::Squash(None) => ":SQUASH".to_string(),
-        Op::Squash(Some(ids)) => {
-            let mut v = ids
-                .iter()
-                .map(|(oid, f)| format!("{}{}", oid.to_string(), spec(*f)))
-                .collect::<Vec<_>>();
-            v.sort();
-            format!(":squash({})", v.join(","))
-        }
-        Op::Linear => ":linear".to_string(),
-        Op::Unsign => ":unsign".to_string(),
-        #[cfg(feature = "incubating")]
-        Op::Adapt(adapter) => format!(":adapt={}", adapter),
-        #[cfg(feature = "incubating")]
-        Op::Link(mode) => format!(":link={}", mode),
-        #[cfg(feature = "incubating")]
-        Op::Export => ":export".to_string(),
-        #[cfg(feature = "incubating")]
-        Op::Unlink => ":unlink".to_string(),
-        Op::Subdir(path) => format!(":/{}", parse::quote_if(&path.to_string_lossy())),
-        Op::File(dest_path, source_path) => {
-            if source_path == dest_path {
-                format!("::{}", parse::quote_if(&dest_path.to_string_lossy()))
-            } else {
-                format!(
-                    "::{}={}",
-                    parse::quote_if(&dest_path.to_string_lossy()),
-                    parse::quote_if(&source_path.to_string_lossy())
-                )
-            }
-        }
-        Op::Prune => ":prune=trivial-merge".to_string(),
-        Op::Prefix(path) => format!(":prefix={}", parse::quote_if(&path.to_string_lossy())),
-        Op::Pattern(pattern) => format!("::{}", parse::quote_if(pattern)),
-        #[cfg(feature = "incubating")]
-        Op::Embed(path) => {
-            format!(":embed={}", parse::quote_if(&path.to_string_lossy()),)
-        }
-        Op::Author(author, email) => {
-            format!(":author={};{}", parse::quote(author), parse::quote(email))
-        }
-        Op::Committer(author, email) => {
-            format!(
-                ":committer={};{}",
-                parse::quote(author),
-                parse::quote(email)
-            )
-        }
-        Op::Message(m, r) if r.as_str() == MESSAGE_MATCH_ALL_REGEX.as_str() => {
-            format!(":{}", parse::quote(m))
-        }
-        Op::Message(m, r) => {
-            format!(":{};{}", parse::quote(m), parse::quote(r.as_str()))
-        }
-        Op::HistoryConcat(r, filter) => {
-            format!(":concat({}{})", r.to_string(), spec(*filter))
-        }
-        #[cfg(feature = "incubating")]
-        Op::Unapply(r, filter) => {
-            format!(":unapply({}{})", r.to_string(), spec(*filter))
-        }
-        Op::Hook(hook) => {
-            format!(":hook={}", parse::quote(hook))
-        }
     }
 }
 
@@ -1870,16 +1609,6 @@ fn pre_process_tree<'a>(
     )?;
 
     Ok(tree)
-}
-
-/// Create a filter that is the result of feeding the output of `first` into `second`
-pub fn chain(first: Filter, second: Filter) -> Filter {
-    opt::optimize(to_filter(Op::Chain(first, second)))
-}
-
-/// Create a filter that is the result of overlaying the output of `first` onto `second`
-pub fn compose(first: Filter, second: Filter) -> Filter {
-    opt::optimize(to_filter(Op::Compose(vec![first, second])))
 }
 
 /// Compute the warnings (filters not matching anything) for the filter applied to the tree
