+++ josh-core/src/filter/opt.rs
@@ -49,7 +49,10 @@ pub fn simplify(filter: Filter) -> Filter {
     if let Some(f) = SIMPLIFIED.lock().unwrap().get(&filter) {
         return *f;
     }
-    rs_tracing::trace_scoped!("simplify", "spec": spec2(&to_op(filter)));
+    rs_tracing::trace_scoped!(
+        "simplify",
+        "spec": crate::flang::spec2(&to_op(filter))
+    );
     let original = filter;
     let result = to_filter(match to_op(filter) {
         Op::Compose(filters) => {
