--- josh-filter/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-authors = ["Christian Schilling <christian.schilling@esrlabs.com>"]
-description = "GIT filter cli"
-edition = "2024"
-keywords = ["git", "monorepo", "workflow", "scm"]
-license-file = "LICENSE"
-name = "josh-filter"
-readme = "../README.md"
-repository = "https://github.com/josh-project/josh"
-version = "22.4.15"
-
-[dependencies]
-josh-core = { path = "../josh-core" }
-josh-graphql = { path = "../josh-graphql" }
-josh-templates = { path = "../josh-templates" }
-env_logger = { workspace = true }
-serde_json = { workspace = true }
-defer = { workspace = true }
-clap = { workspace = true }
-rs_tracing = { workspace = true }
-juniper = { workspace = true }
-git2 = { workspace = true }
-
-[features]
-incubating = []
